fix: rename filenames into kebab case
