+++ electron/binary-downloader-window.ts
@@ -1,7 +1,7 @@
 import { app, BrowserWindow } from "electron";
 import * as path from "path";
 
-import type { ApiResponsesToBinaryDownloader } from "@/@types/response.binaryDownloader";
+import type { ApiResponsesToBinaryDownloader } from "@/@types/response.binary-downloader";
 
 import { baseUrl } from "./context";
