+++ electron/dialog.ts
@@ -9,7 +9,7 @@ import type {
 } from "@/@types/response.controller";
 import type { SpawnResult } from "@/@types/spawn";
 
-import { sendMessageToController } from "./controllerWindow";
+import { sendMessageToController } from "./controller-window";
 import { ffprobePath } from "./ffmpeg";
 import { encodeJson } from "./lib/json";
 import { spawn } from "./lib/spawn";
