+++ electron/ffmpeg-stream/stream.ts
@@ -13,7 +13,7 @@ import type { Readable, Writable } from "stream";
 import { PassThrough } from "stream";
 import { promisify } from "util";
 
-import { sendMessageToController } from "../controllerWindow";
+import { sendMessageToController } from "../controller-window";
 import { ffmpegPath } from "../ffmpeg";
 import { encodeJson } from "../lib/json";
