+++ electron/type-guard.ts
