+++ src/@types/response.binary-downloader.d.ts
