+++ src/controller/movie-picker/local/local-movie-picker.tsx
@@ -6,7 +6,7 @@ import { useCallback, useState } from "react";
 import type { TMovieItemLocal } from "@/@types/queue";
 import type { Movie } from "@/@types/types";
 import { isLoadingAtom, messageAtom } from "@/controller/atoms";
-import { typeGuard } from "@/typeGuard";
+import { typeGuard } from "@/type-guard";
 
 type Props = {
   onChange: (val: TMovieItemLocal | undefined) => void;
