+++ src/controller/queue/comment-item.tsx
@@ -3,9 +3,9 @@ import { useMemo } from "react";
 
 import type { CommentQueue } from "@/@types/queue";
 import { GridDisplay } from "@/controller/display";
-import { StatusDisplay } from "@/controller/queue/StatusDisplay";
+import { StatusDisplay } from "@/controller/queue/status-display";
 
-import Styles from "./ConvertItem.module.scss";
+import Styles from "./convert-item.module.scss";
 
 type Props = {
   queue: CommentQueue;
